Manual running of the dynamic version

+++ Code/DMOP/src/main/java/py/edu/fiuni/dmop/dto/SolutionTraffic.java
@@ -13,9 +13,9 @@ public class SolutionTraffic {
     private Double sloCost;
     private Double licencesCost;
     private Double maxUseLink;
-    private Integer delayCost;
+    private Double delayCost;   // It was defined as int previously
     private Integer distance;
-    private Integer numberInstances;
+    private Double numberInstances; // It was defined as int previously
     private Double throughput;
 
     @Override
